Move the CovPoolTimline component
Move the `CovPoolTimline` component to `components` dir to improve
readability of `withTimeline` hoc.
