+++ solidity-v1/dashboard/src/components/modal/coverage-pools/components/index.js
@@ -1,1 +1,2 @@
 export * from "./WithdrawalOverview"
+export * from "./Timeline"
